Update patch set 4
Patch Set 4:

Patchset #3 adds "findPlugins.sh".  This code isn't used (at least not yet).  It uses the Nexus API to query the repo to see what plugins are available.  It's useful just to do a quick check of what's in the repo.  Possibly we could move to using this to get the plugins that we store into the CM image, instead of having a hard-coded list.   Putting the code here because I did not want to lose the work.

Patch-set: 4
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>