Add unconditional Consul key delete
Add a script to consul-loader-container that
attempts to delete a Consul key and ignores
failures rather than looping indefinitely.
This will allow us to safely add a k8s Job
to each DCAE microservice's Helm chart to delete
the microservice's Consul configuration key when
the microservice is undeployed using Helm.

Issue-ID: DCAEGEN2-2669
Signed-off-by: Jack Lucas <jflos@sonoris.net>
Change-Id: I0e9534c4b488d3a3f7248434712900ed5b593489